Создал класс окна документации, придумал и реализовал навигационную систему. Сделал файлы документации вплоть до Расчёт сил.xml

+++ viewport.h
@@ -58,7 +58,7 @@ public:
     void setPaintVector(QVector<PhObject>& vec);
 
 signals:
-    void whellScrolled(int value);
+    void whellScrolled(double value);
     void camScrolled();
 };
